+++ lib/mebius.dart
@@ -7,6 +7,7 @@ library;
 export 'src/mebius.dart' show Mebius;
 export 'src/mebius_broadcaster.dart' show MebiusBroadcaster;
 export 'src/mebius_client.dart' show MebiusClient;
+export 'src/mebius_delivery.dart' show MebiusDelivery;
 export 'src/mebius_error.dart' show MebiusError, MebiusErrorCode;
 export 'src/mebius_events.dart'
     show